Move the readme and package download code to another class
